+++ src/scripts/modules/views/project.js
@@ -1,121 +1,24 @@
-import storage from '../db/storage';
-import todoController from '../controller/todo';
-import { myCreateElement, generateBtn, openCurrentTabAndContainer } from './helpers';
+import { openCurrentTabAndContainer } from './helpers';
 import todoView from './todo';
+import generatePrContainer from './project_container_generator';
+import generateProjectFooter from './project_footer_generator';
+import generatePrTabs from './project_tabs_generator';
 
 const projectView = (projectController) => {
-  const generateProject = (project) => {
-    const mainRight = document.querySelector('.main__right');
-    if (project == null) { project = projectController.currentProject; }
-    // eslint-disable-next-line no-use-before-define
-    mainRight.appendChild(generatePrContainer(project));
-  };
-
   const renderProjects = () => {
     document.querySelector('.project-tabs').innerHTML = '';
     document.querySelector('.main__right').innerHTML = '';
     // eslint-disable-next-line no-use-before-define
-    generatePrTabs();
+    generatePrTabs(projectController, generateProject);
   };
 
-  const generateProjectFooter = (project) => {
-    const prFooter = document.createElement('div');
-
-    const addTodoBtn = generateBtn('Add Todo');
-    addTodoBtn.classList.add('add-todo-btn');
-    const addTodoForm = todoView(project).generateForm(null, 'add-todo-form');
-    const prRmBtn = generateBtn('Remove Project');
-    prRmBtn.classList.add('project__rmbtn');
-
-    addTodoBtn.addEventListener('click', (e) => {
-      e.target.textContent = e.target.textContent === 'Add Todo' ? 'Close' : 'Add Todo';
-      e.target.nextElementSibling.classList.toggle('open');
-    });
-    addTodoForm.addEventListener('submit', e => {
-      e.preventDefault();
-      const { elements } = e.target;
-      const title = elements.todoTitle.value;
-      const description = elements.todoDesc.value;
-      const dueDate = elements.todoDue.value;
-      const priority = elements.todoPriority.value;
-      todoController(project).createTodo(title, description, dueDate, priority);
-      todoView(project).clearTodoField(elements);
-      generateProject(project);
-      renderProjects();
-      openCurrentTabAndContainer(project);
-    });
-    prRmBtn.addEventListener('click', () => {
-      const nextProject = projectController.removeProject(project.id);
-      projectController.currentProject = nextProject;
-      generateProject(nextProject);
-      renderProjects();
-      openCurrentTabAndContainer(nextProject);
-    });
-
-    prFooter.appendChild(addTodoBtn);
-    prFooter.appendChild(addTodoForm);
-    prFooter.appendChild(prRmBtn);
-    return prFooter;
-  };
-
-  const openTab = (project, tabItem) => {
-    document.querySelectorAll('.project-tabs__item').forEach(tab => {
-      tab.classList.remove('active');
-    });
-    document.querySelectorAll('.project-container').forEach(item => {
-      item.classList.remove('active');
-    });
-    document.getElementById(project.id).classList.add('active');
-    tabItem.classList.add('active');
-  };
-
-  const generatePrTabs = () => {
-    const tabsContainer = document.querySelector('.project-tabs');
-    generateProject(projectController.currentProject);
-
-    if (storage.projects.length > 0) {
-      storage.projects.forEach((pro) => {
-        const tabItem = myCreateElement('div', 'project-tabs__item');
-        tabItem.textContent = pro.title;
-        tabItem.setAttribute('data-id', pro.id);
-        tabsContainer.appendChild(tabItem);
-        tabItem.addEventListener('click', (e) => {
-          projectController.currentProject = projectController.getProject(e.target.getAttribute('data-id'));
-          generateProject(projectController.currentProject);
-          openTab(projectController.currentProject, e.target);
-        });
-      });
-    }
-  };
-
-  const generatePrContainer = (project) => {
-    const prContainer = myCreateElement('div', 'project-container');
-    prContainer.setAttribute('id', project.id);
-
-    const pr = myCreateElement('div', 'project');
-    const prItem = myCreateElement('div', 'project__item');
-
-    const prTitle = myCreateElement('h2', 'project__title');
-    prTitle.textContent = project.title;
-
-    const prDesc = myCreateElement('p', 'project__desc');
-    prDesc.textContent = project.description;
-
-    const prTodosItem = myCreateElement('div', 'project__todos');
-    prContainer.appendChild(pr);
-    pr.appendChild(prItem);
-    prItem.appendChild(prTitle);
-    prItem.appendChild(prDesc);
-    prItem.appendChild(prTodosItem);
-
-    if (project.todos.length > 0) {
-      project.todos.forEach(todo => {
-        prTodosItem.appendChild(todoView(project).generateTodo(todo, project, renderProjects));
-      });
-    }
-    prTodosItem.appendChild(generateProjectFooter(project, projectController));
-
-    return prContainer;
+  const generateProject = (project) => {
+    const mainRight = document.querySelector('.main__right');
+    if (project == null) { project = projectController.currentProject; }
+    mainRight.appendChild(generatePrContainer(
+      project, projectController, todoView(project).generateTodo,
+      generateProjectFooter, renderProjects, generateProject,
+    ));
   };
 
   const clearPrField = (el) => {
